:recycle: Split structs
